backend/requirements.txt
asgiref==3.6.0
atomicwrites==1.4.1
attrs==23.1.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.0.12
colorama==0.4.6
cryptography==41.0.2
defusedxml==0.7.1
Django==3.2.3
django-cors-headers==3.13.0
django-filter==2.4.0
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-simplejwt==5.2.2
djoser==2.2.0
drf-extra-fields==3.7.0
drf-spectacular==0.25.1
exceptiongroup==1.1.1
filetype==1.2.0
flake8==6.0.0
flake8-isort==6.0.0
idna==3.4
inflection==0.5.1
iniconfig==2.0.0
gunicorn==20.0.4
isort==5.12.0
jsonschema==4.17.3
mccabe==0.7.0
oauthlib==3.2.2
packaging==23.1
Pillow==10.0.0
pluggy==0.13.1
psycopg2-binary==2.9.3
py==1.11.0
pycodestyle==2.10.0
pycparser==2.21
pyflakes==3.0.1
PyJWT==2.1.0
pyrsistent==0.19.3
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3
PyYAML==6.0
requests==2.26.0
requests-oauthlib==1.3.1
sentry-sdk==1.16.0
social-auth-app-django==5.2.0
social-auth-core==4.4.2
sqlparse==0.4.4
toml==0.10.2
tomli==2.0.1
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.14
